+++ mantener/django_portfolio/settings.py
@@ -37,8 +37,7 @@ INSTALLED_APPS = [
     "django.contrib.sessions",
     "django.contrib.messages",
     "django.contrib.staticfiles",
-    "main",
-    "portfolio"
+    "main"
 ]
 
 MIDDLEWARE = [
