+++ mantener/django_portfolio/urls.py
@@ -22,14 +22,14 @@ from django.conf import settings
 
 # importamos en donde se encuentran las nuevas ligas a utilizar
 from main import views
-from portfolio.views import render_proyects2
+
 
 urlpatterns = [
     path("admin/", admin.site.urls),
     # Agregamos los patrones de las ligas que hemos creado
     path('', views.home, name='home'), #con el nombre es como colocarle un ID a la liga
     path('projects/', include('portfolio.urls')), # con esta ruta renderizamos los proyectos
-    path('projects2/', render_proyects2, name='projects2')
+    path('projects2/', views.render_proyects2, name='projects2')
 ]
 
 urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
