+++ mantener/main/admin.py
@@ -1,6 +1,7 @@
 from django.contrib import admin
 
 # Importamos el modulo para que pueda leer la clase creada
-from .models import Main
+from .models import Main, Portfolio
 
 admin.site.register(Main)
+admin.site.register(Portfolio)
