+++ mantener/main/views.py
@@ -1,6 +1,13 @@
-from django.shortcuts import render, get_object_or_404
-from .models import Portfolio
+from django.shortcuts import render
 
+# Importamos la consulta a la base de datos
+from .models import Main, Portfolio
+
+# Create your views here.
+def home(request):
+    '''Funcion para que nos arroje la pagina de inicio'''
+    main = Main.objects.all()
+    return render(request, 'home.html', {'main':main})
 
 
 # Create your views here.
