--- requirements.txt
@@ -1,13 +1,0 @@
-asgiref==3.6.0
-black==23.1.0
-click==8.1.3
-colorama==0.4.6
-Django==4.1.7
-mypy-extensions==1.0.0
-packaging==23.0
-pathspec==0.11.0
-Pillow==9.4.0
-platformdirs==3.1.1
-sqlparse==0.4.3
-tomli==2.0.1
-tzdata==2022.7
